Remove an unused check when loading dictionary
If there was no learnedDict in storage, `chrome.storage.local.get()`
will return `{}`, so `result.learnedDict` will be `undefined`, which
will be correctly ignored by `Object.assign` (and `if (result) {...}`
would have executed anyway). The current check would work if `result`
itself is somehow `undefined`, but in that case we might want to fail
loudly anyway.